pyproject.toml
[project]
name = "fastloom"
version = "0.3.3"
description = "Core package"
authors = []
readme = "README.md"

[[tool.poetry.source]]
name = "nexus-registry"
url = "https://nexus.bluprint.ir/repository/group-package-registry/simple/"
priority = "primary"

[tool.mypy]
plugins = ["pydantic.mypy"]
[[tool.mypy.overrides]]
module = "aredis_om"
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
jdatetime = "^4.1.1"
beanie = "^2.0.0"
python-jose = "^3.5.0"
pydantic = { version = "^2.11", extras = ["email"] }
uvicorn = "^0.35.0"
orjson = "3.10.14"
pyyaml = "^6.0.2"

fastapi = { version = "^0", optional = true }
python-multipart = { version = "*", optional = true }
faststream = { version = "^0.5.28", extras = [
    "rabbit",
    "otel",
], optional = true }
celery = { version = "^5.5.3", optional = true }
httpx = { version = "^0.28.0", optional = true }
requests = { version = "^2.32.0", optional = true }
redis-om = {git = "https://github.com/redis/redis-om-python", rev = "main", optional = true}
openai = { version = "^1", optional = true }
logfire = { version = "^4.0.0", extras = ["system-metrics", "asgi", "fastapi", "celery", "httpx", "requests", "redis", "pymongo", "openai"] }
sentry-sdk = { extras = ["fastapi"], version = "^2.0.0" }
opentelemetry-distro = "*"
opentelemetry-exporter-otlp = "*"
opentelemetry-instrumentation-aio-pika = { version = "*", optional = true }
babel = "^2.17.0"
jinja2 = "^3.1.6"

# dev
pre-commit = { version = "^4.2.0", optional = true }
ruff = { version = "^0.12.5", optional = true }
mypy = { version = "^1.17.0", optional = true }
ipykernel = { version = "^6.30.0", optional = true }

[tool.poetry.extras]
FastAPI = [
    "fastapi",
    "python-multipart",
    "uvicorn",
]
Rabbit = [
    "fastapi",
    "uvicorn",
    "faststream",
    "opentelemetry-instrumentation-aio-pika",
]
Celery = ["celery"]
HTTPx = ["httpx"]
Requests = ["requests"]
Redis = ["redis-om"]
Mongodb = ["beanie"]
OpenAI = ["openai"]
Dev = ["pre-commit", "ruff", "mypy", "ipykernel"]

[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.21.1"
deepdiff = "^6.7.0"
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
freezegun = "^1.2.2"
pytest-cov = "^4.1.0"
pytest-xdist = "^3.5.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
types-pyyaml = "^6.0.12.20240917"
types-requests = "^2.32.4.20250809"
types-redis = "^4.6.0.20241004"
types-python-jose = "^3.5.0.20250531"

[project.scripts]
launch = "fastloom.launcher.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
